Switch role of parameter and filter context
We need to add a flag to the filter context to mark when requested query
parameters are empty after filter matching. This is needed to break
filtering (or at least have a marker for later) when a filter contains
restricted values. If this is the case, an empty response can be
generated (it can be treated as a non-existent parameter value) as all
query parameters serve as AND filter.