Create empty/ExceptionReport response when query gets invalid
When filtering restricted query parameters there might be cases where
every parameter value is restricted. In those cases the proxy emulates
the behaviour of the backend without delegating the request, i.e. a
response is created which would be returned by the backend if it was
requested with an unknown parameter. This makes it transparent when a
user has no access to specific parameters but requesting them anyway.